requirements.txt
langchain>=0.3.0,<0.4.0
langgraph>=0.6.2,<0.7.0
langgraph-cli>=0.3.6,<0.4.0
langgraph-sdk>=0.2.0,<0.3.0
langgraph-checkpoint-sqlite>=2.0.11,<3.0.0
langsmith>=0.4.8,<0.5.0
langchain-community>=0.3.27,<0.4.0
langchain-core>=0.3.72,<0.4.0
langchain-openai>=0.3.28,<0.4.0
langchain-experimental>=0.3.4,<0.4.0
notebook>=7.4.4,<8.0.0
python-dotenv>=1.1.1,<2.0.0
langchain-chroma>=0.2.5,<0.3.0
openevals>=0.1.0,<0.2.0
openai>=1.98.0,<2.0.0
ipython>=9.4.0,<10.0.0
pyppeteer>=2.0.0,<3.0.0
langgraph-supervisor>=0.0.29,<0.0.30
langgraph-cli[inmem]>=0.3.6,<0.4.0
ipykernel>=6.30.0,<7.0.0
pydantic[email]>=2.11.7,<3.0.0
docx2txt>=0.9,<1.0.0
pypdf>=5.9.0,<6.0.0
pymupdf>=1.26.3,<2.0.0
streamlit>=1.28.0,<2.0.0
python-multipart>=0.0.20
fastapi>=0.116.1